@@ -29,31 +29,30 @@ gem 'omniauth-google-oauth2'
 gem 'pg'
 gem 'phony_rails'
 gem 'rack-cors', require: 'rack/cors'
-gem 'rails', '4.2.8'
-gem 'rails-i18n', '~> 4.0.0'
-gem 'rails-observers'
-gem 'rake'
+gem 'rails', '~> 5.1'
+gem 'rails-i18n'
+#gem 'rails-observers'
 gem 'readthis'
-gem 'redis', '>= 3.2.0', require: ['redis', 'redis/connection/hiredis']
+gem 'redis', require: ['redis', 'redis/connection/hiredis']
 gem 'remotipart', '~> 1.2' # [Q] Are we using this?
 gem 'selectize-rails' # why not npm?
-gem 'slim-rails', '~> 3.1.1'
+gem 'slim-rails'
 
-# Use Paper Trail for containing a full history of our edits.
+## Use Paper Trail for containing a full history of our edits.
 gem 'action_parameter'
 gem 'aws-sdk', '~> 2'
-gem 'paperclip', '~> 5.0.0'
+gem 'paperclip'
 gem 'paper_trail'
 gem 'rmagick' # rmagick for image processing
-# AWS SDK for Rails - makes SES integration easy
+## AWS SDK for Rails - makes SES integration easy
 gem 'actionkit_connector', git: 'https://github.com/SumOfUs/actionkit_connector', branch: 'master'
 gem 'airbrake', '~> 5.7.1'
 gem 'airbrake-ruby', '1.7.1'
 gem 'aws-sdk-rails'
 gem 'bootstrap-sass', '~> 3.3.5'
 gem 'browser', '~> 2.0', '>= 2.0.3'
-gem 'coffee-rails', '~> 4.1.0'
-gem 'compass-rails', '~> 3.0.2' # was using git master branch before
+#gem 'coffee-rails', '~> 4.1.0'
+gem 'compass-rails' # was using git master branch before
 gem 'config'
 gem 'envyable', require: 'envyable/rails-now'
 gem 'friendly_id'
