@@ -3,7 +3,7 @@ module Api
   module Stateless
     module GoCardless
       class PaymentMethodsController < StatelessController
-        before_filter :authenticate_request!
+        before_action :authenticate_request!
 
         def index
           @payment_methods = PaymentHelper::GoCardless.payment_methods_for_member(@current_member).active
