@@ -3,7 +3,7 @@ module Api
   module Stateless
     module GoCardless
       class TransactionsController < StatelessController
-        before_filter :authenticate_request!
+        before_action :authenticate_request!
 
         def index
           @transactions = PaymentHelper::GoCardless.transactions_for_member(@current_member)
