@@ -2,9 +2,9 @@
 require 'share_progress'
 
 class Share::SharesController < ApplicationController
-  before_filter :set_resource
-  before_filter :find_page
-  before_filter :authenticate_user!
+  before_action :set_resource
+  before_action :find_page
+  before_action :authenticate_user!
 
   def new
     @share = share_class.new(new_defaults)
