@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class PaymentController < ApplicationController
-  skip_before_action :verify_authenticity_token
+  skip_before_action :verify_authenticity_token, raise: false
   before_action :localize_from_page_id, only: :transaction
 
   def transaction
